1. В IFan организовал расчет рабочих точек для первоначальной кривой (свойство OriginalCurve) и новой кривой (свойство NewCurve). 2. Количество точек задаётся вручную в коде с помощью свойства [IFan, строка #428] CountArray (установлено = 8); 2. Вывел точки NewCurve на печать.

+++ Libraries/Description_of_objects/FanData.cs
@@ -54,12 +54,12 @@ public record FanData
     /// <summary>
     ///     Минимальный объем воздуха, [м3/ч]
     /// </summary>
-    public required int MinVolumeFlow { get; init; }
+    public required double MinVolumeFlow { get; init; }
 
     /// <summary>
     ///     Максимальный объем воздуха, [м3/ч]
     /// </summary>
-    public required int MaxVolumeFlow { get; init; }
+    public required double MaxVolumeFlow { get; init; }
 
     /// <summary>
     ///     Коэффиуиенты полинома 6-й степени Pv(Q) - полного давления от объемного воздуха
